+++ packages/visual-grid-client/index.d.ts
@@ -143,7 +143,7 @@ declare namespace Eyes {
        * The accessibility level to use for the screenshots
        * Default value: 'None'
        */
-      accessibilityLevel?:AccessibilityLevel
+      accessibilitySettings?:AccessibilitySettings
 
       /**
        * DEPRECATED, use batchNotify.
@@ -189,7 +189,13 @@ declare namespace Eyes {
     }
   }
 
-  type AccessibilityLevel = "None"|"AA"|"AAA"
+  type AccessibilityLevel = "AA"|"AAA"
+  type AccessibilityGuidelinesVersion = "WCAG_2_0"|"WCAG_2_1"
+
+  interface AccessibilitySettings {
+    level: AccessibilityLevel
+    guidelinesVersion: AccessibilityGuidelinesVersion
+  }
 
   export namespace Check {
     interface Options {
@@ -273,12 +279,6 @@ declare namespace Eyes {
        * and not for normal production use.
        */
       sendDom?:boolean
-
-      /**
-       * The accessibility level to use for the screenshot
-       * Default value: 'None'
-       */
-      accessibilityLevel?:AccessibilityLevel
     }
 
     interface Selector {