+++ packages/visual-grid-client/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- support accessibilitySettings
 
 
 ## 14.0.2 - 2020/5/18